refactor(ProfilePopup): вынести вспомогательные компоненты полей в отдельные файлы
