+++ src/components/pages/Profile/ProfilePopup/fields/ValidityInput.js
@@ -1,0 +1,17 @@
+import React, { PureComponent } from 'react';
+import MaskedInput from 'react-text-mask';
+
+class ValidityInput extends PureComponent {
+   render() {
+      return (
+         <MaskedInput 
+            { ...this.props.input }
+            placeholder="00/00" 
+            guide={ false }
+            mask={[/\d/, /\d/, '/', /\d/, /\d/]}
+         />
+      );
+   }
+};
+
+export default ValidityInput;
